@@ -1,13 +1,84 @@
 {
   "name": "react-app",
-  "version": "0.1.0",
   "private": true,
+  "version": "0.1.0",
   "sideEffects": false,
   "scripts": {
+    "prebuild": "start-server-and-test 3000 cypress && npm run sitemap",
     "build": "craco build",
     "start": "react-scripts start",
     "test": "react-scripts test",
-    "eject": "react-scripts eject"
+    "cypress": "cypress run --headless",
+    "eject": "react-scripts eject",
+    "sitemap": "node scripts/sitemap.js",
+    "locale": "npm run prebuild && node scripts/locale.js",
+    "lighthouse": "start-server-and-test 3000 'lighthouse --output=json --output-path=reports/lighthouse/index.json http://localhost:3000'"
+  },
+  "dependencies": {
+    "@emotion/react": "^11.11.3",
+    "@emotion/styled": "^11.11.0",
+    "@mui/icons-material": "^5.15.5",
+    "@mui/material": "^5.15.5",
+    "axios": "^1.6.5",
+    "firebase": "^10.7.2",
+    "i18next": "^23.7.18",
+    "i18next-browser-languagedetector": "^7.2.0",
+    "i18next-http-backend": "^2.4.2",
+    "nanoid": "^5.0.4",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-helmet": "^6.1.0",
+    "react-i18next": "^14.0.1",
+    "react-markdown": "^9.0.1",
+    "react-router-dom": "^6.21.3",
+    "react-scripts": "5.0.1",
+    "typescript": "^5.3.3",
+    "web-vitals": "^3.5.1"
+  },
+  "devDependencies": {
+    "@craco/craco": "^7.1.0",
+    "@principalstudio/html-webpack-inject-preload": "^1.2.7",
+    "@testing-library/jest-dom": "^6.2.0",
+    "@testing-library/react": "^14.1.2",
+    "@testing-library/user-event": "^14.5.2",
+    "@tsconfig/cypress": "^1.0.1",
+    "@types/jest": "^29.5.11",
+    "@types/node": "^20.11.5",
+    "@types/react": "^18.2.48",
+    "@types/react-dom": "^18.2.18",
+    "@types/react-helmet": "^6.1.11",
+    "@typescript-eslint/eslint-plugin": "^6.19.0",
+    "@typescript-eslint/parser": "^6.19.0",
+    "compression-webpack-plugin": "^11.0.0",
+    "css-minimizer-webpack-plugin": "^6.0.0",
+    "cypress": "^13.6.3",
+    "dotenv": "^16.3.2",
+    "dotenv-webpack": "^8.0.1",
+    "ejs-loader": "^0.5.0",
+    "eslint": "^8.56.0",
+    "eslint-config-airbnb": "^19.0.4",
+    "eslint-config-prettier": "^9.1.0",
+    "eslint-config-react-app": "^7.0.1",
+    "eslint-plugin-prettier": "^5.1.3",
+    "eslint-plugin-unused-imports": "^3.0.0",
+    "express": "^4.18.2",
+    "glob": "^10.3.10",
+    "html-critical-webpack-plugin": "^2.1.0",
+    "html-webpack-plugin": "^5.6.0",
+    "image-minimizer-webpack-plugin": "^4.0.0",
+    "lighthouse": "^11.4.0",
+    "mini-css-extract-plugin": "^2.7.7",
+    "openai": "^4.25.0",
+    "prettier": "^3.2.4",
+    "purgecss-webpack-plugin": "^5.0.0",
+    "sass": "^1.70.0",
+    "start-server-and-test": "^2.0.3",
+    "stylelint": "^16.2.0",
+    "stylelint-scss": "^6.1.0",
+    "terser-webpack-plugin": "^5.3.10",
+    "webpack-bundle-analyzer": "^4.10.1",
+    "yargs": "^17.7.2",
+    "zlib": "^1.0.5"
   },
   "eslintConfig": {
     "extends": [
