@@ -1,9 +1,10 @@
-import React from "react";
-import { render, screen } from "@testing-library/react";
+/* eslint-disable import/no-extraneous-dependencies */
+import { render } from "@testing-library/react";
 import App from "./App";
 
-test("renders learn react link", () => {
-  render(<App />);
-  const linkElement = screen.getByText(/learn react/i);
-  expect(linkElement).toBeInTheDocument();
+describe("<App />", () => {
+	it("renders without crashing", () => {
+		const { container } = render(<App />);
+		expect(container).toBeTruthy();
+	});
 });
