@@ -1,22 +1,30 @@
+import React from "react";
 import Box from "@mui/material/Box";
 import Button from "@mui/material/Button";
 import Dialog from "@mui/material/Dialog";
 import DialogContent from "@mui/material/DialogContent";
 import Typography from "@mui/material/Typography";
 import { useTranslation } from "react-i18next";
-import { Link, useNavigate } from "react-router-dom";
+import { useNavigate, Link } from "react-router-dom";
 import AchivementIcon from "../../../components/Icons/Achivement";
-import { languages } from "../../../context/Language/Language";
-import { Pages } from "../../../utils/config";
+import { languages } from "../../../i18n";
+import { Pages } from "../../../services/config";
 import "./page.scss";
 
 function Congratulations() {
 	const navigate = useNavigate();
+	const buttonRef = React.useRef<HTMLButtonElement>(null);
 	const { t, i18n } = useTranslation(["subscribe", "common"]);
 	const currentLanguage = languages.find((lang) => lang.id === i18n.language);
 
 	const objSpacing = 2;
 	const recommendationsPagePath = `${currentLanguage?.path}${Pages.recommendations}`;
+
+	React.useEffect(() => {
+		if (buttonRef.current) {
+			buttonRef.current.focus();
+		}
+	}, []);
 
 	return (
 		<Dialog
@@ -54,17 +62,21 @@ function Congratulations() {
 					letterSpacing={0}
 					className="CongratulationMessage"
 				>
-					You've subscribed! Visit our website to explore the best jobs in the Web3 and A.I. space.
+					{t("messages.youveSubscribed")}
 				</Typography>
 				<Button
 					fullWidth
+					autoFocus
+					focusRipple
+					tabIndex={0}
+					type="button"
 					component={Link}
 					variant="contained"
 					sx={{ zIndex: 9999 }}
 					to={recommendationsPagePath}
 					className="GetStartedButton"
 				>
-					Get Started Here
+					{t("cta.getStartedHere")}
 				</Button>
 				<Typography
 					component="p"
@@ -73,7 +85,7 @@ function Congratulations() {
 					textAlign="center"
 					className="CongratulationNotice"
 				>
-					*Please check your email to confirm your subscription.
+					*{t("messages.pleaseCheckYourEmail")}
 				</Typography>
 			</DialogContent>
 			<div className="Fireworks">
