@@ -1,15 +1,22 @@
-import React, { Suspense } from "react";
+/* eslint-disable no-console */
+import React from "react";
 import Skeleton from "@mui/material/Skeleton";
-
+import { useNavigate, useLocation } from "react-router-dom";
 import SwitcherButton from "./SwitcherButton/SwitcherButton";
-import { Language, useLanguage, languages } from "../../context/Language/Language";
+import { languages, getCurrentLanguage, changeLanguage, getPreviewsLanguage, Language } from "../../i18n";
 
 const LanguagePopup = React.lazy(() => import(/* webpackChunkName: 'LanguagePopup' */ "./LanguagePopup/LanguagePopup"));
 
 function LanguageSwitcher() {
-	const { language, changeLanguage } = useLanguage();
+	const location = useLocation();
+	const navigate = useNavigate();
 	const [anchorEl, setAnchorEl] = React.useState<null | HTMLElement>(null);
+
 	const open = Boolean(anchorEl);
+	const hashFragment = location.hash;
+	const searchParams = location.search;
+	const currentPath = location.pathname;
+	const pathSegments = currentPath.split("/").filter((p) => p);
 
 	const handleClick = (event: React.MouseEvent<HTMLElement>) => {
 		setAnchorEl(event.currentTarget);
@@ -17,20 +24,33 @@ function LanguageSwitcher() {
 	const handleClose = () => {
 		setAnchorEl(null);
 	};
+
 	const selecteLanguage = (language: Language) => {
-		changeLanguage(language);
-		setAnchorEl(null);
+		const lastLanguage = getPreviewsLanguage();
+		let newPath = `${language.path}/${pathSegments.join("/")}${searchParams}${hashFragment}`;
+		const currentPath = `${location.pathname}${searchParams}${hashFragment}`;
+
+		changeLanguage(language.value);
+
+		if (lastLanguage?.path) {
+			newPath = newPath.replace(lastLanguage.path, "");
+		}
+
+		if (currentPath !== newPath) {
+			navigate(newPath);
+			setAnchorEl(null);
+		}
 	};
 
 	return (
 		<>
 			<SwitcherButton
 				open={open}
-				language={language}
 				onClick={handleClick}
+				language={languages.find((l) => l.path === `/${pathSegments[0]}`) || getCurrentLanguage()}
 			/>
 			{open && (
-				<Suspense
+				<React.Suspense
 					fallback={
 						<Skeleton
 							width={384}
@@ -42,12 +62,12 @@ function LanguageSwitcher() {
 					<LanguagePopup
 						open={open}
 						anchorEl={anchorEl}
-						language={language}
+						language={languages.find((l) => l.value === getCurrentLanguage().value)}
 						languages={languages}
 						onClose={handleClose}
 						selecteLanguage={selecteLanguage}
 					/>
-				</Suspense>
+				</React.Suspense>
 			)}
 		</>
 	);
